+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## ChangeLog
 
+### 2.x.x
+* Fixed the issue #6
+* Changed the function name of `selectRange` to `selectNextElement`
+
 ### 2.0.16
 * Deleted a variable never used
 * Re-define the parameters for `filterInputValue` function